+++ src/main/java/blackjack/domain/player/Profit.java
@@ -1,4 +1,4 @@
-package blackjack.domain;
+package blackjack.domain.player;
 
 import java.util.Objects;
