+++ src/test/java/blackjack/domain/GameManagerTest.java
@@ -3,6 +3,7 @@ package blackjack.domain;
 import blackjack.domain.card.Card;
 import blackjack.domain.card.CardNumber;
 import blackjack.domain.card.CardShape;
+import blackjack.domain.game.GameManager;
 import blackjack.domain.player.Dealer;
 import blackjack.domain.player.Gambler;
 import org.junit.jupiter.api.DisplayName;
